Add universal flow fields, vivid multi-zone color, and performance overhaul
- Universal flow system: 7 pre-computed velocity fields (radial, rotate, swirl,
  bubble, ring, vortex, vertical) with semi-Lagrangian advection at 0.8px/step
- All presets ship with strong flow defaults (5 active channels each)
- Vivid palettes: full-gamut cosine colors with 2nd/3rd harmonics for depth
- Spatial color offset creates distinct color zones across organism
- GS contrast stretch (V*3.0) reveals sharp RD detail
- SmoothLife contrast curve (world^1.8) prevents blown-out look
- Thickness slider for line dilation control
- Pre-computed noise pool (6 full-res frames, zero per-frame cost)
- Float32 throughout all engines for ~2x speedup
- Lenia 8→18 FPS, SmoothLife 6→12, MNCA 8→15, GS 33→55

+++ plugins/cellular_automata/engine_base.py
@@ -17,7 +17,7 @@ class CAEngine(ABC):
 
     def __init__(self, size=512):
         self.size = size
-        self.world = np.zeros((size, size), dtype=np.float64)
+        self.world = np.zeros((size, size), dtype=np.float32)
         self.generation = 0
 
     @abstractmethod
